back/djangoAPI/requirements.txt
gunicorn==20.0.4
asgiref==3.2.10
Django==3.0.9
django-cors-headers==3.4.0
djangorestframework==3.11.0
psycopg2==2.8.4
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0